Java

+++ syntax/Math.java
@@ -1,6 +1,6 @@
 public class Math{
 
-	public static int add(int a, int b){
+	public static double add(double a, double b){
 		return a + b;
 	}
